Got my data in pretty good. Now I just need to make some UI.

+++ HoloCOVIDUnity/Assets/Data/CovidDayData.cs
@@ -1,6 +1,6 @@
 ﻿public struct CovidDayData
 {
-    public int Confirmed { get; set; }
-    public int Deaths { get; set; }
-    public int Recovered { get; set; }
+    public float Confirmed { get; set; }
+    public float Deaths { get; set; }
+    public float Recovered { get; set; }
 }
